Исправлен тип переменной для хранения client_id

+++ steam_web/Auth/v2/DTO/CAuthentication_GetAuthSessionInfo_Request.cs
@@ -5,6 +5,6 @@ namespace SteamWeb.Auth.v2.DTO;
 [ProtoContract]
 public class CAuthentication_GetAuthSessionInfo_Request
 {
-    [ProtoMember(1)] public uint client_id { get; set; }
+    [ProtoMember(1)] public ulong client_id { get; set; }
 }
